Initiated tests

+++ config/initializers/yandex_newsletter.rb
@@ -1,2 +1,4 @@
-Que.execute 'DELETE FROM que_jobs'
-YandexNewsPuller.enqueue
+unless Rails.env.test?
+  Que.execute 'DELETE FROM que_jobs'
+  YandexNewsPuller.enqueue
+end
